+++ src/day07.rs
@@ -14,20 +14,20 @@ pub struct Day07 {
 }
 
 impl Problem for Day07 {
-    type Output = i32;
+    type Output = i64;
 
-    fn part_one(&self) -> i32 {
+    fn part_one(&self) -> i64 {
         self.solve(&mut [0, 1, 2, 3, 4])
     }
 
-    fn part_two(&self) -> i32 {
+    fn part_two(&self) -> i64 {
         self.solve(&mut [5, 6, 7, 8, 9])
     }
 }
 
 impl Day07 {
-    fn solve(&self, phases: &mut [i32; 5]) -> i32 {
-        let program: Vec<i32> = fs::read_to_string(&self.input)
+    fn solve(&self, phases: &mut [i64; 5]) -> i64 {
+        let program: Vec<i64> = fs::read_to_string(&self.input)
             .unwrap()
             .trim()
             .split(',')
@@ -47,7 +47,7 @@ impl Day07 {
     }
 }
 
-fn run(program: &[i32], phases: &[i32]) -> i32 {
+fn run(program: &[i64], phases: &[i64]) -> i64 {
     let mut intcodes: Vec<_> = phases.iter().map(|phase| init(&program, *phase)).collect();
 
     let mut signal = 0;
@@ -64,7 +64,7 @@ fn run(program: &[i32], phases: &[i32]) -> i32 {
     signal
 }
 
-fn init(program: &[i32], phase: i32) -> Intcode {
+fn init(program: &[i64], phase: i64) -> Intcode {
     let mut intcode = Intcode::new(program.to_vec());
     let output = intcode.run(&[phase]);
